requirements.txt
# Web framework
fastapi==0.115.11
pydantic==2.10.6
uvicorn==0.34.0

# AI/ML libraries
haystack-ai==2.7.0

# PyTorch with CPU support
torch==2.6.0
torchvision==0.21.0

# Utilities
python-dotenv==1.0.1